Create Gallery branch with all pre-PR changes
This branch is rebased against Google's Ampstart master, and contains all the gallery-specific code from the latest `next`. It also re-enables the validate task, re-structures the Gallery code to fit Sriram's guidelines, removes a lot of custom CSS in favour of using BassCSS classes, modifies the validate task in order to only check amp.html files, re-enables the cssnano task, removes all the parallax code and <script> tags that failed to pass validation.

This branch is intended to be used as a base when creating PRs for Gallery -- start with this branch, move to another branch, reset this commit, and add only what you need into the PR.

I'm mostly just making this because I'm terrified of accidentally deleting it, and I _really_ don't want to do all that re-factoring again :)